Move CloudFileStore to useCloudFile
